% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCI_var.R
\name{BGPsv}
\alias{BGPsv}
\title{CI calculation for the total variance parameter in a single-factor random-effects ANOVA model}
\usage{
BGPsv(Y, A, alpha = 0.1, Nmax = 1e+05)
}
\arguments{
\item{Y}{observed outcomes in an ANOVA model}

\item{A}{factor level}

\item{alpha}{desired significance level for the confidence interval}

\item{Nmax}{total number of generalized pivotal statistics to be generated}
}
\value{
\describe{
  \item{GCI}{ generalized CI from Park and Burdick (2004) }
  \item{BCI}{ asymptotic CI from Burdick and Graybill (1984) }
}
}
\description{
Consider a random-effects ANOVA model, \eqn{Y_{ij}=\mu+u_i+\epsilon_{ij}}, where the random-effects \eqn{u_i\sim N(0,\sigma_a^2)} and
 the random error components \eqn{\epsilon_{ij}\sim N(0,\sigma^2)}. We want to compute CI for the total variance parameter \eqn{\sigma_a^2+\sigma^2}.
An asymptotic approach and a generalized pivotal test approach are implemented. See the two references.
}
\references{
Burdick, R.K., Graybill, F.A., 1984. Confidence intervals on linear combinations of variance components in the unbalanced one-way classification. Technometrics 26, 131–136.

Park, D.J., Burdick, R.K., 2004. Confidence intervals on total variance in a regression model with an unbalanced one-fold nested error structure. Comm. Statist. Theory Methods 33, 2735–2743.
}
