% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCI_var.R
\name{GCIsv}
\alias{GCIsv}
\title{Generalized confidence interval for the total variance in a single-factor random-effects ANOVA model}
\usage{
GCIsv(Y, A, alpha = 0.1, Nmax = 2000)
}
\arguments{
\item{Y}{observed outcomes in an ANOVA model}

\item{A}{factor level}

\item{alpha}{desired significance level for the confidence interval}

\item{Nmax}{total number of generalized pivotal statistics to be generated}
}
\value{
\describe{
  \item{GCI}{ generalized CI }
  \item{Qt}{ generated generalized pivotal statistics for the total variance }
}
}
\description{
Consider a random-effects ANOVA model, \eqn{Y_{ij}=\mu+u_i+\epsilon_{ij}}, where the random-effects \eqn{u_i\sim N(0,\sigma_a^2)} and
 the random error components \eqn{\epsilon_{ij}\sim N(0,\sigma^2)}. We want to compute CI for the total variance parameter \eqn{\sigma_a^2+\sigma^2}.
We adopt the generalized pivotal test approach to computing the CI, based on
sequentially ``calculating" the within-group variance parameter \eqn{\sigma^2} and the between-group variance parameter \eqn{\sigma_a^2}.
}
\examples{
A = rep(1:10, times=5:14)
s2a = 0.05; s2 = 0.95
Y = rnorm(10)[A]*sqrt(s2a) + rnorm(length(A))*sqrt(s2)
BGPsv(Y,A)
aa = GCIsv(Y,A); aa$GCI
}
\references{
Bai,X., Bai,Y. and Wu,B. (2019) Generalized confidence interval calculation for the total variance in a single-factor random-effects ANOVA model
            with application to medical device comparison problems. tech report.
}
