% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTukey.R
\name{GTukey}
\alias{GTukey}
\title{A generalized Tukey method for multiple pairwise comparisons in ANOVA model}
\usage{
GTukey(nr, alpha = 0.05, W = NULL)
}
\arguments{
\item{nr}{the sample sizes for each factor level.}

\item{alpha}{overall significance level}

\item{W}{the relative type I errors assigned to each comparison.}
}
\value{
\describe{
  \item{Bonf}{ critial values computed from Bonferroni correction }
  \item{Tukey}{ Tukey HSD approximate value (common to all) }
  \item{WT}{ weighted Tukey HSD critical values }
  \item{FP}{ overall familywise type I errors }
  \item{FPi}{ individual test type I errors }
}
}
\description{
Given an ANOVA model with r levels, we are interested in testing the pairwise mean differences (1 vs 2,...; 2 vs 3,... etc). 
We consider a weighted multiple hypothesis testing problem, where we allow different tests
to get unequal type I errors, and in the meantime, we want to control the overall type I errors at a pre-specified level.
}
\examples{
GTukey(c(15,10,5), W=3:1)
}
\references{
X,X., and Wu,B. (2019) A generalized Tukey method for multiple pairwise comparisons in ANOVA model.
}
