% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTukey.R
\name{PGTukey}
\alias{PGTukey}
\title{Multiple testing adjusted p-value (familywise error rate) in a weighted post hoc comparisons in an ANOVA model.}
\usage{
PGTukey(nr, Qx = 2, ik = 1, W = NULL)
}
\arguments{
\item{nr}{the sample sizes for each factor level.}

\item{Qx}{observed t-statistic for a comparison}

\item{ik}{index of the comparison that we want to compute the multiple testing adjusted p-value}

\item{W}{the relative type I errors assigned to each comparison.}
}
\value{
\describe{
  \item{alpha}{ multiple testing adjusted p-value (FWER) }
  \item{FPi}{ individual test type I errors }
}
}
\description{
Given an ANOVA model with r levels, we are interested in testing the pairwise mean differences (1 vs 2,...; 2 vs 3,... etc). 
We consider a weighted multiple hypothesis testing problem, where we allow different tests
to get unequal type I errors, and in the meantime, we want to control the overall type I errors (familywise error rate; FWER)
at a pre-specified level. For each comparison, we compute the multiple testing adjusted p-value, which
is the smallest FWER that will lead to a significant test.
}
\examples{
GTukey(c(15,10,5), W=3:1)
}
\references{
Ling,S., Johnson,C., Ellermann,J., Eberly,L. and Wu,B. (2019) A generalized Tukey method for weighted multiple pairwise comparisons in an ANOVA model.
}
